Inline ReflectiveClassInspector and move off of immutable nodes API
