--- interop/kotlinx-metadata/classinspectors/reflect/gradle.properties
@@ -1,4 +1,0 @@
-POM_ARTIFACT_ID=kotlinpoet-classinspector-reflective
-POM_NAME=KotlinPoet (Reflective ClassInspector implementation)
-POM_DESCRIPTION=Reflective ClassInspector implementation.
-POM_PACKAGING=jar
