+++ settings.gradle.kts
@@ -24,7 +24,6 @@ include(
     ":kotlinpoet",
     ":interop:javapoet",
     ":interop:kotlinx-metadata:classinspectors:elements",
-    ":interop:kotlinx-metadata:classinspectors:reflect",
     ":interop:kotlinx-metadata:core",
     ":interop:kotlinx-metadata:specs",
     ":interop:kotlinx-metadata:specs-tests",
